+++ polls/tests/test_views.py
@@ -1,0 +1,87 @@
+import datetime
+
+from django.http import HttpResponseRedirect
+from django.test import TestCase, Client
+from django.urls import reverse
+from django.utils import timezone
+
+from polls.models import Question
+
+
+def create_question(question_text, days):
+    """Create a question for polls app.
+
+    Create a question with the given `question_text` and published the
+    given number of `days` offset to now (negative for questions published
+    in the past, positive for questions that have yet to be published).
+    """
+    time = timezone.now() + datetime.timedelta(days=days)
+    return Question.objects.create(question_text=question_text, pub_date=time)
+
+
+class QuestionIndexViewTest(TestCase):
+    """Test question index page."""
+
+    def setUp(self):
+        """Set up a client to request later."""
+        self.client = Client()
+
+    def test_no_question(self):
+        """If no question exist, an approprite message is displayed."""
+        response = self.client.get(reverse('polls:index'))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_past_question(self):
+        """Questions with a pub_date in the past are displayed on the index page."""
+        create_question(question_text="Past question.", days=-30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_future_question(self):
+        """Questions with a pub_date in the future aren't displayed on the index page."""
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_future_question_and_past_question(self):
+        """Even if both past and future questions exist, only past questions are displayed."""
+        create_question(question_text="Past question.", days=-30)
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_two_past_questions(self):
+        """The questions index page may display multiple questions."""
+        create_question(question_text="Past question 1.", days=-30)
+        create_question(question_text="Past question 2.", days=-5)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question 2.>', '<Question: Past question 1.>']
+        )
+
+class QuestionDetailViewTests(TestCase):
+    """Test question detail page."""
+
+    def test_future_question(self):
+        """The detail view of a question with a pub_date in the future returns a 404 not found."""
+        future_question = create_question(question_text='Future question.', days=5)
+        url = HttpResponseRedirect(reverse('polls:detail', args=(future_question.id,)))
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 404)
+
+    def test_past_question(self):
+        """The detail view of a question with a pub_date in the past displays the question's text."""
+        past_question = create_question(question_text='Past Question.', days=-5)
+        url = reverse('polls:detail', args=(past_question.id,))
+        response = self.client.get(url)
+        self.assertContains(response, past_question.question_text)
